+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v0.2.0-dev (2015-XX-XX)
+* Enhancements
+  * Additional custom stylesheet for admin. 
+    SCSS for this is stored in `web/static/css/brando.custom.scss`.
 
 ## v0.1.0 (2015-09-28)