+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.2.0-dev (2015-XX-XX)
 * Enhancements
-  * Additional custom stylesheet for admin. 
-    SCSS for this is stored in `web/static/css/brando.custom.scss`.
+  * Additional custom stylesheet option for admin. 
+    SCSS for this is stored in `web/static/css/brando.custom.scss`, and compiled to `priv/static/css/brando.custom.css`.
   * Optional `admin_hostname` plug to check `conn.host` has `admin` prefix. 
     This is to make sure the admin area is only accessed from `admin.myapp.com`