+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     SCSS for this is stored in `web/static/css/brando.custom.scss`, and compiled to `priv/static/css/brando.custom.css`.
   * Optional `admin_hostname` plug to check `conn.host` has `admin` prefix. 
     This is to make sure the admin area is only accessed from `admin.myapp.com`
+  * Added forced re-rendering of all pages through Villain parser.
 
 * Bug fixes
   * Fixed css bugs on system info page.